Strip off ANSI colored maven output from deps lines
This extra output got included in the filename and breaks
installation of jars on platforms where the output gets included.
This includes at least Linux, but I have been unable to reproduce
on MacOS so far. This is a stopgap until I can figure out why this
output is happening and fully disable it.

The regexp matching in this method could also be improved to avoid
breaking on future format changes.

Fixes jruby/jar-dependencies#92.

See also jruby/jruby#8606.